+++ sdk-extensions/incubator/src/test/java/io/opentelemetry/sdk/extension/incubator/fileconfig/component/TestAuthComponentProvider.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.sdk.extension.incubator.fileconfig.component;
+
+import io.opentelemetry.api.incubator.authenticator.ExporterAuthenticator;
+import io.opentelemetry.api.incubator.config.DeclarativeConfigProperties;
+import io.opentelemetry.sdk.autoconfigure.spi.internal.ComponentProvider;
+import java.util.Collections;
+import java.util.Map;
+
+public class TestAuthComponentProvider implements ComponentProvider<ExporterAuthenticator> {
+  @Override
+  public Class<ExporterAuthenticator> getType() {
+    return ExporterAuthenticator.class;
+  }
+
+  @Override
+  public String getName() {
+    return "test_auth";
+  }
+
+  @Override
+  public ExporterAuthenticator create(
+      DeclarativeConfigProperties config, ComponentProviderLoader componentProviderLoader) {
+    return new ExporterAuthenticator() {
+      ;
+
+      @Override
+      public String getName() {
+        return "test_auth";
+      }
+
+      @Override
+      public Map<String, String> getAuthenticationHeaders() {
+        return Collections.singletonMap("auth_provider_key1", "value1");
+      }
+    };
+  }
+}
